+++ src/axis/YAxis.js
@@ -1,8 +1,28 @@
-const Axis = require('./Axis')
 
-class YAxis extends Axis {
+const defaults = {
+  stroke: '#d7d5d2',
+  'stroke-width': 1
+}
+
+class YAxis {
   constructor(obj = {}, world) {
-    super(obj, world);
+    this.world = world
+    this.attrs = Object.assign({}, defaults, obj)
+  }
+  ticks(n = 6) {
+    n = n === 0 ? 0 : n - 1
+    let scale = this.world.yScale.scale
+    let max = this.world.yScale.max
+    let ticks = []
+    for (let i = 0; i <= n; i += 1) {
+      let dec = i / n
+      let num = dec * max
+      ticks.push({
+        pos: scale(num),
+        label: parseInt(num, 10)
+      })
+    }
+    return ticks
   }
   drawTicks(x) {
     return this.ticks().map((o) => {
@@ -13,7 +33,8 @@ class YAxis extends Axis {
   }
   build() {
     let attrs = this.attrs
-    let range = this.world.xScale.scale.range()
+    let domain = this.world.xScale.scale.domain()
+    console.log(domain)
     let x = 0
     attrs = Object.keys(attrs).map((k) => {
       return `${k}="${attrs[k]}"`
@@ -21,7 +42,7 @@ class YAxis extends Axis {
     let ticks = this.drawTicks(x)
     return `<g>
       ${ticks}
-      <line x1="${x}" y1="${range[0]}" x2="${x}" y2="${range[1]}" ${attrs}/>
+      <line x1="${x}" y1="${domain[0]}" x2="${x}" y2="${domain[1]}" ${attrs}/>
     </g>`
   }
 }
