+++ src/shapes/Line.js
@@ -4,6 +4,8 @@ const Shape = require('./Shape')
 const defaults = {
   fill: 'none',
   stroke: 'steelblue',
+  'stroke-width': 4,
+  'stroke-linecap': 'round'
 }
 
 class Line extends Shape {
